% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{caterpillar_plot}
\alias{caterpillar_plot}
\title{caterpillar_plot}
\usage{
caterpillar_plot(obj, ci = 0.9, col_id = TRUE, ...)
}
\arguments{
\item{obj}{An object of type \code{ss_ranef}.}

\item{ci}{The width of the credible interval that should be used. Defaults to 0.9.}

\item{col_id}{Whether the plot should color in points by their unique identifier.}

\item{...}{Currently not in use.}
}
\description{
Create a caterpillar plot of the random effects produced by \code{ss_ranef} models.
}
