% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{coef.ivd}
\alias{coef.ivd}
\title{Extract coefficients from a fitted ivd model}
\usage{
\method{coef}{ivd}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ivd}.}

\item{...}{Not used.}
}
\value{
Named numeric vector of posterior means.
}
\description{
Extract coefficients
}
\details{
Convenience alias for \code{\link[=fixef]{fixef()}}: posterior means of the fixed location
and scale effects.
}
\author{
Philippe Rast
}
