% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.autocorrelation_fft}
\alias{.autocorrelation_fft}
\title{Use fast Fourier transform to compute ACF}
\usage{
.autocorrelation_fft(chain)
}
\value{
acf
}
\description{
Fast Fourier transform algorithm to compute the ACF across the whole chain lengt.
As noted by Vehtari et al. (2021), Section 3.2
}
\author{
Philippe Rast
}
\keyword{internal}
