% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.extract_to_mcmc}
\alias{.extract_to_mcmc}
\title{Extract samples to mcmc object}
\usage{
.extract_to_mcmc(obj)
}
\value{
mcmc object
}
\description{
Extract samples to mcmc object
}
\author{
Philippe Rast
}
\keyword{internal}
