% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{.pooled_draws}
\alias{.pooled_draws}
\title{Pool the post-warmup draws of all chains into one matrix.}
\usage{
.pooled_draws(object)
}
\arguments{
\item{object}{ivd object}
}
\value{
Matrix of pooled draws (iterations x parameters).
}
\description{
Pool the post-warmup draws of all chains into one matrix.
}
\keyword{internal}
