% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.require_suggest}
\alias{.require_suggest}
\title{Helper to check for suggested package}
\usage{
.require_suggest(pkg, feature)
}
\arguments{
\item{pkg}{requested package}

\item{feature}{requested feature}
}
\description{
Helper to check for suggested package
}
\author{
philippe
}
