% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{funnel_plot}
\alias{funnel_plot}
\title{funnel_plot}
\usage{
funnel_plot(obj, ...)
}
\arguments{
\item{obj}{An object fit with \code{ss_ranef_mv()}}

\item{...}{Currently not in use}
}
\description{
funnel_plot
}
