% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_mv_data.R
\name{gen_mv_data}
\alias{gen_mv_data}
\title{gen_mv_data}
\usage{
gen_mv_data(n, n_j)
}
\arguments{
\item{n}{A numeric specifying the number of units.}

\item{n_j}{A numeric specifying the number of observations per unit.}
}
\description{
gen_mv_data
}
