% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivd.R
\name{ivd_alpha}
\alias{ivd_alpha}
\title{ivd_alpha}
\usage{
ivd_alpha(
  y,
  unit,
  burnin = 1000,
  iter = 1000,
  chains = 4,
  priors = NULL,
  vars2monitor = c("alpha", "gamma", "sigma", "tau", "theta")
)
}
\arguments{
\item{y}{A vector containing the outcome}

\item{unit}{A vector of the same length as \code{y} containing a unique identifier}

\item{burnin}{The number of iterations to use as burnin. Defaults to 1000.}

\item{iter}{The number of iterations to use for estimating the parameters. Defaults to 1000.}

\item{chains}{The number of MCMC chains to use. Defaults to 4.}

\item{priors}{A named list to specify priors.  Defaults to NULL. See README for details.}

\item{vars2monitor}{A vector containing the names of which parameters to monitor. See details below.}
}
\value{
An object of type \code{ivd}.
}
\description{
\code{ivd_alpha()} fits a random intercepts model with a spike-and-slab prior on the random effects
}
\details{
The parameters that can be tracked are:
\itemize{
\item alpha: The fixed effect for the intercept
\item gamma: The inclusion indicators for the random effects
\item sigma: The residual standard deviation
\item tau: The standard deviation of the random effects
\item theta: The random effects
}
}
