% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivd.R
\name{ivd_beta}
\alias{ivd_beta}
\title{ivd_beta}
\usage{
ivd_beta(
  y,
  X,
  unit,
  burnin = 1000,
  iter = 1000,
  chains = 4,
  priors = NULL,
  vars2monitor = c("alpha", "beta", "gamma", "rho", "sigma", "tau1", "tau2", "theta1",
    "theta2")
)
}
\arguments{
\item{y}{A vector containing the outcome}

\item{X}{A vector containing the predictor}

\item{unit}{A vector of the same length as \code{y} containing a unique identifier}

\item{burnin}{The number of iterations to use as burnin. Defaults to 1000.}

\item{iter}{The number of iterations to use for estimating the parameters. Defaults to 1000.}

\item{chains}{The number of MCMC chains to use. Defaults to 4.}

\item{priors}{A named list to specify priors.  Defaults to NULL. See README for details.}

\item{vars2monitor}{A vector containing the names of which parameters to monitor. See details below.}
}
\value{
An object of type \code{ivd}.
}
\description{
\code{ivd_beta()} fits a mixed-effects model with
random intercepts and random slopes, with a spike-and-slab prior on the random slopes.
}
\details{
The parameters that can be tracked are:
\itemize{
\item alpha: The fixed effect for the intercept
\item beta: The fixed effect for the slope
\item gamma: The inclusion indicators for the slope random effects
\item rho: The correlation between theta1 and theta2
\item sigma: The residual standard deviation
\item tau1: The standard deviation for alpha
\item tau2: The standard deviation for beta
\item theta1: The random effects for alpha
\item theta2: The random effects for beta
}
}
