% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check.R
\name{pp_check}
\alias{pp_check}
\title{Posterior predictive check}
\usage{
pp_check(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Passed to methods.}
}
\value{
See the method documentation.
}
\description{
Posterior predictive check
}
\details{
Generic; see \code{\link[=pp_check.ivd]{pp_check.ivd()}}. Defined in-package so that \code{ivd} does not
depend on \pkg{bayesplot}; if bayesplot is attached its same-named
generic dispatches to the ivd method just the same.
}
\author{
Philippe Rast
}
