% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prepare_data_for_nimble}
\alias{prepare_data_for_nimble}
\title{Define data from formula}
\usage{
prepare_data_for_nimble(data, location_formula, scale_formula)
}
\arguments{
\item{data}{Data object in long format}

\item{location_formula}{Formula for location}

\item{scale_formula}{Formula for scale}
}
\description{
Define data from formula
}
