% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ivd.R
\name{print.ivd_sim}
\alias{print.ivd_sim}
\title{Print an ivd_sim object}
\usage{
\method{print}{ivd_sim}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ivd_sim}.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for simulated ivd data
}
\author{
Philippe Rast
}
