% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip_diagnostics.R
\name{print.pip_diagnostics}
\alias{print.pip_diagnostics}
\title{Print PIP Monte Carlo diagnostics}
\usage{
\method{print}{pip_diagnostics}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pip_diagnostics}.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{pip_diagnostics} objects
}
\author{
Philippe Rast
}
