% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip_fdr.R
\name{print.pip_fdr}
\alias{print.pip_fdr}
\title{Print a pip_fdr selection}
\usage{
\method{print}{pip_fdr}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pip_fdr}.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{pip_fdr} objects
}
\author{
Philippe Rast
}
