% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.posterior_summary}
\alias{print.posterior_summary}
\title{print.posterior_summary}
\usage{
\method{print}{posterior_summary}(x, ...)
}
\arguments{
\item{x}{An object of type \code{ivd}}

\item{...}{Currently not in use}
}
\description{
Print method for \code{posterior_summary()}
}
