% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.ranef_summary}
\alias{print.ranef_summary}
\title{print.ranef_summary}
\usage{
\method{print}{ranef_summary}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{Currently not in use}
}
\description{
Print method for \code{ranef_summary()}
}
