% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.ivd}
\alias{print.summary.ivd}
\title{Print method for summary.ivd objects}
\usage{
\method{print}{summary.ivd}(x, ...)
}
\arguments{
\item{x}{A \code{summary.ivd} object.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Print the posterior summary assembled by \code{\link[=summary.ivd]{summary.ivd()}}
}
\author{
Philippe Rast
}
