% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{ranef_summary}
\alias{ranef_summary}
\title{ranef_summary}
\usage{
ranef_summary(obj, ci = 0.9, as_df = FALSE, digits = 2)
}
\arguments{
\item{obj}{An object of type \code{ivd}}

\item{ci}{The width of the credible interval that should be used. Defaults to 0.9.}

\item{as_df}{Whether a \code{data.frame} should instead be returned. Defaults to FALSE.}

\item{digits}{The number of digits to which the output should be rounded. Defaults to 2.}
}
\description{
Summarizes the posterior distribution of the random effects contained in \code{ivd} objects
}
