% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivd.R
\name{build_ivd_model}
\alias{build_ivd_model}
\title{Build and compile NIMBLE model and MCMC once
This functin is exported for use in 'future'}
\usage{
build_ivd_model(code, constants, dummy_data, dummy_inits, useWAIC = TRUE)
}
\description{
Build and compile NIMBLE model and MCMC once
This functin is exported for use in 'future'
}
