% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{codaplot}
\alias{codaplot}
\title{Traceplot from the coda package}
\usage{
codaplot(obj, parameters = NULL, type = "traceplot")
}
\arguments{
\item{obj}{ivd object}

\item{parameters}{Provide parameters of interest as c("parameter1", "paramter2") etc.}

\item{type}{Coda plot. Defaults to 'traceplot'}
}
\value{
Specified coda plot
}
\description{
For more plots see coda
}
\author{
Philippe Rast
}
