% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{codaplot}
\alias{codaplot}
\title{Traceplot from the coda package}
\usage{
codaplot(obj, parameters = NULL, type = "traceplot", askNewPage = TRUE)
}
\arguments{
\item{obj}{ivd object}

\item{parameters}{Provide parameters of interest using names from the summary() output (e.g., "Intc", "scl_Intc", "sd_Intc", "R\[scl_Intc, Intc\]", "pip\[Intc, 5\]"). Defaults to NULL (plots all parameters).}

\item{type}{Coda plot. Defaults to 'traceplot'. See coda for more options such as 'acfplot', 'densplot' etc.}

\item{askNewPage}{Should user be prompted for next plot. Defaults to \code{TRUE}}
}
\value{
Specified coda plot
}
\description{
For more plots see coda
}
\author{
Philippe Rast
}
