% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{.summary_table}
\alias{.summary_table}
\title{Filter coda object}
\usage{
.summary_table(stats, Kr)
}
\arguments{
\item{stats}{summary object}

\item{Kr}{Number of random location efx}
}
\value{
filtered coda
}
\description{
Filter coda object
}
\author{
philippe
}
\keyword{internal}
