% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivd.R
\name{ivd}
\alias{ivd}
\title{ivd}
\usage{
ivd(location_formula, scale_formula, data, niter, nburnin = NULL, ...)
}
\arguments{
\item{location_formula}{lme4 type formula type}

\item{scale_formula}{lme4 type formula type}

\item{data}{Data in long format}

\item{niter}{Number of iterations after burnin}

\item{nburnin}{Number of burnin. Defaults to the same amount as niter (i.e,. with niter = 5000, nburnin is also 5000).}

\item{...}{not defined}
}
\value{
An object of type \code{ivd}.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Philippe Rast
}
