% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivd.R
\name{ivd}
\alias{ivd}
\title{ivd}
\usage{
ivd(location_formula, scale_formula, data, niter)
}
\arguments{
\item{data}{list of data objects}

\item{groups}{Number of groups (clusters)}

\item{group_id}{Vector of length N with sequentially ordered group id's}
}
\value{
An object of type \code{ivd}.
}
\description{
ivd
}
\details{
The parameters that can be tracked are:
\itemize{
\item alpha: The fixed effect for the intercept
\item beta: The fixed effect for the slope
\item gamma: The inclusion indicators for the slope random effects
\item rho: The correlation between theta1 and theta2
\item sigma: The residual standard deviation
\item tau1: The standard deviation for alpha
\item tau2: The standard deviation for beta
\item theta1: The random effects for alpha
\item theta2: The random effects for beta
}
}
