% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivd.R
\name{ivd}
\alias{ivd}
\title{Main function to set up and run parallel MCMC using nimble and future}
\usage{
ivd(location_formula, scale_formula, data, niter, nburnin = NULL, ...)
}
\arguments{
\item{location_formula}{A formula for the location model}

\item{scale_formula}{A formula for the scale model}

\item{data}{Data frame in long format for analysis}

\item{niter}{Total number of MCMC iterations after burnin}

\item{nburnin}{Number of burnin iterations, defaults to the same as niter}
}
\description{
Main function to set up and run parallel MCMC using nimble and future
}
