% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ivd}
\alias{plot.ivd}
\title{Plot method for ivd objects}
\usage{
\method{plot}{ivd}(
  x,
  type = "pip",
  pip_level = 0.75,
  variable = NULL,
  col_id = TRUE,
  legend = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of type \code{ivd}.}

\item{type}{Defaults to 'pip', other options are 'funnel' and 'outcome'.}

\item{pip_level}{Defines a value for the posterior inclusion probability. Defaults to 0.75.}

\item{variable}{Name of a specific variable. Defaults to \code{NULL}}

\item{col_id}{Whether the plot should color in points by their unique identifier.}

\item{legend}{Should legend be included? Defaults to \code{TRUE}.}

\item{...}{Currently not in use.}
}
\description{
Plot method for ivd objects
}
\author{
Philippe Rast
}
