% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivd.R
\name{run_MCMC_allcode}
\alias{run_MCMC_allcode}
\title{` Create a function to be loaded on each worker. This needs to be exported for future to be able to load it.}
\usage{
run_MCMC_allcode(
  seed,
  data,
  constants,
  code,
  niter,
  nburnin,
  useWAIC = TRUE,
  inits,
  additional_monitors = "NULL"
)
}
\description{
` Create a function to be loaded on each worker. This needs to be exported for future to be able to load it.
}
