% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivd.R
\name{run_MCMC_allcode}
\alias{run_MCMC_allcode}
\title{` Create a function with all the needed code}
\usage{
run_MCMC_allcode(
  seed,
  data,
  constants,
  code,
  niter = 5000,
  useWAIC = TRUE,
  inits
)
}
\description{
` Create a function with all the needed code
}
