% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivd.R
\name{run_MCMC_compiled_model}
\alias{run_MCMC_compiled_model}
\title{Run MCMC on an already compiled model}
\usage{
run_MCMC_compiled_model(
  compiled,
  seed,
  new_data,
  new_inits,
  niter,
  nburnin,
  useWAIC = TRUE,
  ...
)
}
\description{
Run MCMC on an already compiled model
}
