% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ivd}
\alias{summary.ivd}
\title{Summary of posterior samples}
\usage{
\method{summary}{ivd}(object, digits = 3, pip = "all", ...)
}
\arguments{
\item{object}{ivd object}

\item{digits}{Integer (Default: 2, optional). Number of digits to round to when printing.}

\item{pip}{Print pip and model parameters ('all'); Only pip ('pip'), or only model parameeters 9('model'). Defaults to 'all'}

\item{...}{Not used}
}
\value{
summary.ivd object
}
\description{
Summarize ivd object
}
\author{
Philippe Rast
}
